правки по замечаниям
я переписала страницу вывода одного лота. постаралась добавить проверки на целое положительное число. подготовленные выражения пока не знаю, как применять.
