+++ templates/error.php
@@ -10,6 +10,10 @@
     </nav>
     <section class="lot-item container">
         <h2>404 Страница не найдена</h2>
-        <p>Данной страницы не существует на сайте.</p>
+        <?php if (isset($errors)): ?>
+             <?php foreach ($errors as $val): ?>
+                <p><?= $val; ?></p>
+             <?php endforeach; ?>
+     <?php endif; ?>
     </section>
 </main>
